Detailed instructions for Babel presets
Include an example of specifying babel presets from within the webpack configuration, since it's not obvious how to do this if you're not using a separate babel config. Linked to babelrc doc and 'Setting up Babel 6' article since the syntax shown in the previously linked article for including query parameters doesn't work in this case as there are multiple loaders used.